Split out FDisableLinkFilter
