+++ src/js/Content/Features/Community/CCommunityBase.js
@@ -1,6 +1,7 @@
 import {Context, ContextType} from "../../modulesContent";
 import FHideTrademarks from "../Common/FHideTrademarks";
 import FBackToTop from "../Common/FBackToTop";
+import FDisableLinkFilter from "../Common/FDisableLinkFilter";
 import FConfirmDeleteComment from "./FConfirmDeleteComment";
 import FHideSpamComments from "./FHideSpamComments";
 import FFavoriteEmoticons from "./FFavoriteEmoticons";
@@ -12,6 +13,7 @@ export class CCommunityBase extends Context {
         features.push(
             FHideTrademarks,
             FBackToTop,
+            FDisableLinkFilter,
             FConfirmDeleteComment,
             FHideSpamComments,
             FFavoriteEmoticons,
